Move internal code to impl; API tests to docs
